+++ .github/workflows/types.yml
@@ -3,7 +3,7 @@ name: TypeScript Check
 on:
   pull_request:
     paths:
-      - '*.d.ts'
+      - '**.ts'
       - package.json
       - .github/workflows/types.yml
 
@@ -14,8 +14,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v1
 
-      - name: Install TypeScript
-        run: npm install -g typescript@latest
+      - run: npm install
 
       - name: Check typings file
-        run: tsc -p test/types/tsconfig.json
+        run: npx tsd
