--- test/types/tsconfig.json
@@ -1,22 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2018",
-		"lib": [
-			"es2018"
-		],
-		"noEmit": true, /* Do not emit outputs. */
-		"allowSyntheticDefaultImports": true,
-		/* Strict Type-Checking Options */
-		"strict": true /* Enable all strict type-checking options. */,
-		/* Additional Checks */
-		"noUnusedLocals": true /* Report errors on unused locals. */,
-		"noUnusedParameters": true /* Report errors on unused parameters. */,
-		"noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
-		"noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
-		/* Module Resolution Options */
-		"moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */,
-	},
-	"files": [
-		"test.ts"
-	]
-}
